+++ Doolda/Tests/Dummies/DummyError.swift
@@ -1,0 +1,13 @@
+//
+//  DummyError.swift
+//  EditPageUseCaseTest
+//
+//  Created by 김민주 on 2021/11/30.
+//
+
+import Foundation
+
+enum DummyError: Error {
+    case notImplemented
+    case failed
+}
