+++ Doolda/Tests/GetMyIdUseCaseTest/GetMyIdUseCaseTest.swift
@@ -1,0 +1,32 @@
+//
+//  GetMyIdUseCaseTest.swift
+//  GetMyIdUseCaseTest
+//
+//  Created by 김민주 on 2021/11/30.
+//
+
+import XCTest
+
+class GetMyIdUseCaseTest: XCTestCase {
+
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+    }
+
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+
+    func testExample() throws {
+        // This is an example of a functional test case.
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+
+    func testPerformanceExample() throws {
+        // This is an example of a performance test case.
+        measure {
+            // Put the code you want to measure the time of here.
+        }
+    }
+
+}
